+++ packages/client/common/components/card/index.tsx
@@ -11,12 +11,14 @@ type Props = {
 function CardInner({title, children, className, innerRef}: Props) {
     return (
         <div
-            ref={innerRef}
-            className={classNames(['rounded-lg bg-white shadow-lg md:shadow-xl relative overflow-hidden', className])}
+            className={classNames([
+                'rounded-lg bg-white dark:bg-gray-800 shadow-lg md:shadow-xl relative overflow-hidden',
+                className,
+            ])}
         >
             <div className="px-3 pt-2 pb-3 md:p-5 relative box-border w-full">
-                <h4 className="font-bold text-xl mb-2">{title}</h4>
-                {children}
+                <h4 className="font-bold text-xl mb-2 dark:text-white">{title}</h4>
+                <div className="dark:text-white text-gray-700">{children}</div>
             </div>
         </div>
     );