+++ packages/client/common/components/select/index.tsx
@@ -13,12 +13,15 @@ export default function Select(props: Props) {
 
     return (
         <div className="w-full">
-            <label className="block uppercase tracking-wide text-gray-700 text-xs font-bold mb-1" htmlFor={name}>
+            <label
+                className="block uppercase tracking-wide text-gray-700 dark:text-white text-xs font-bold mb-1"
+                htmlFor={name}
+            >
                 {label}
             </label>
             <div className="relative">
                 <select
-                    className="block appearance-none w-full bg-gray-200 border text-gray-700 py-2 px-2 md:px-3 md:py-2 rounded leading-tight focus:outline-none focus:shadow-outline focus:bg-white focus:border-gray-500 "
+                    className="block appearance-none w-full bg-gray-200 dark:bg-gray-500 text-gray-700 dark:text-white py-2 px-2 md:px-3 md:py-2 dark:focus:bg-gray-500 rounded leading-tight focus:outline-none focus:shadow-outline focus:bg-white focus:border-gray-500 "
                     id={name}
                     name={name}
                     value={value}
@@ -30,7 +33,7 @@ export default function Select(props: Props) {
                         </option>
                     ))}
                 </select>
-                <div className="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700">
+                <div className="pointer-events-none absolute inset-y-0 right-0 flex items-center px-2 text-gray-700 dark:text-white">
                     <svg className="fill-current h-4 w-4" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 20 20">
                         <path d="M9.293 12.95l.707.707L15.657 8l-1.414-1.414L10 10.828 5.757 6.586 4.343 8z" />
                     </svg>