+++ packages/client/common/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@threequackinn/client-common-v1.2.1](https://github.com/TheThreeQuackInn/site/compare/@threequackinn/client-common-v1.2.0...@threequackinn/client-common-v1.2.1) (2020-11-22)
+
+
+### Bug Fixes
+
+* **toggle:** add accessibilitu to toggle component ([2923337](https://github.com/TheThreeQuackInn/site/commit/2923337bf3d87f96524aaa7f3161a6311e51b9c6))
+
+# [@threequackinn/client-common-v1.2.0](https://github.com/TheThreeQuackInn/site/compare/@threequackinn/client-common-v1.1.0...@threequackinn/client-common-v1.2.0) (2020-11-22)
+
+
+### Features
+
+* **darkmode:** add darkmode toggle ([b3f3ef6](https://github.com/TheThreeQuackInn/site/commit/b3f3ef604012aca238db92fce270c454a9b8c772))
+
 # [@threequackinn/client-common-v1.1.0](https://github.com/TheThreeQuackInn/site/compare/@threequackinn/client-common-v1.0.0...@threequackinn/client-common-v1.1.0) (2020-10-25)